pyproject.toml
[project]
name = "nightcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.4",
    "async-lru>=2.0.5",
    "asyncpg>=0.30.0",
    "chat-exporter>=2.8.4",
    "discord-py",
    "httpx>=0.28.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.42",
]

[dependency-groups]
dev = ["ruff>=0.12.7"]

[tool.uv.sources]
discord-py = { git = "https://github.com/Rapptz/discord.py.git" }

# ================================= RUFF
[tool.ruff]
# Line length according to PEP 8 - 79 characters
line-length = 79

# Target Python version
target-version = "py310"

exclude = [".venv", "venv", "__pycache__", "migrations", "build", "dist"]

[tool.ruff.lint]
# Enable rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]

# Ignore rules
ignore = ["D202", "D212", "D104", "D101", "D107", "RUF001"]

# Allow unused variables if they start with an underscore
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Docstring handling
[tool.ruff.lint.pydocstyle]
convention = "google"

# Settings for isort
[tool.ruff.lint.isort]
known-first-party = ["src"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

# Formatting options
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
